fixed dbt ephemeral nodes lin
